requirements.txt
django-rest-framework==0.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
Django==4.2.3
django-cors-headers==4.1.0
django-environ==0.10.0
